Move auth0 builder to module
